# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Ejemplo",
        "description": "Documentación de la API generada con Swagger en Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/weather": {
            "get": {
                "tags": [
                    "Clima"
                ],
                "summary": "Obtener el clima por ciudad",
                "operationId": "68be5d208ae4f75b329bcb8e4c648f57",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Nombre de la ciudad (opcional, por defecto 'El Vigia')",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "El Vigia"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del clima obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "city": {
                                            "type": "string",
                                            "example": "El Vigia"
                                        },
                                        "localtime": {
                                            "type": "string",
                                            "example": "2025-05-14 15:00"
                                        },
                                        "temperature_celsius": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 25.3
                                        },
                                        "condition": {
                                            "type": "string",
                                            "example": "Soleado"
                                        },
                                        "humidity": {
                                            "type": "integer",
                                            "example": 40
                                        },
                                        "wind_kph": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 15.2
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron datos para la ciudad especificada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No se encontraron datos para la ciudad especificada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No se pudo obtener el clima."
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/mark-favorite": {
            "post": {
                "tags": [
                    "Favoritos"
                ],
                "summary": "Marcar una ciudad como favorita",
                "operationId": "5dbb6e39046b460c63de4dc76ec5d8be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city"
                                ],
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "example": "El Vigia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ciudad marcada como favorita",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ciudad marcada como favorita."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error por ciudad no encontrada en historial o ya marcada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No puedes marcar como favorita una ciudad que no has buscado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/search-history": {
            "get": {
                "tags": [
                    "Historial"
                ],
                "summary": "Historial de búsquedas del usuario",
                "operationId": "656b500b4b268e1b207bd71c815d723f",
                "responses": {
                    "200": {
                        "description": "Lista del historial",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/favorite-cities": {
            "get": {
                "tags": [
                    "Favoritos"
                ],
                "summary": "Obtener ciudades favoritas del usuario",
                "operationId": "2e866094e3f57cbae0fb147e0a19c024",
                "responses": {
                    "200": {
                        "description": "Lista de ciudades favoritas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Autenticación con token Bearer. Ejemplo: Bearer {token}"
            }
        }
    }
}